Added MISC_MEM and OP_IMM decodings
Added (and completed) MISC_MEM and OP_IMM decodings for the following instructions:
  - ADDI
  - SLTI
  - SLTIU
  - ANDI
  - ORI
  - XORI
  - SLLI
  - SRLI
  - SRAI

Began work on triadic (3-op) OP instructions. Transitioned the code to non-blocking assignments in the entirety of the clocked circuitry. Preliminary coding of FENCE/FENCE.I is that of a NOP from an execution perspective to avoid introduction of pipeline incongruities.